refactor!: Rename libvcs.parse -> libvcs.url
