+++ docs/index.md
@@ -9,7 +9,7 @@
 :hidden:
 
 quickstart
-parse/index
+url/index
 cmd/index
 projects/index
 ```
