--- docs/parse/base.md
@@ -1,9 +1,0 @@
-(parser-framework)=
-
-# Framework: Add and extend URL parsers - `libvcs.parse.base`
-
-```{eval-rst}
-.. automodule:: libvcs.parse.base
-   :members:
-   :undoc-members:
-```
