--- docs/parse/hg.md
@@ -1,9 +1,0 @@
-# Mercurial URL Parser - `libvcs.parse.hg`
-
-For hg, aka `hg(1)`.
-
-```{eval-rst}
-.. automodule:: libvcs.parse.hg
-   :members:
-   :undoc-members:
-```
