+++ docs/url/git.md
@@ -6,12 +6,13 @@ myst:
     "property=og:locale": "en_US"
 ---
 
-# Git URL Parser - `libvcs.parse.git`
+# Git URL Parser - `libvcs.url.git`
 
-Detect, parse, and change git URLs using libvcs's URL parser for `git(1)`. It builds on top of the {ref}`VCS-friendly URL parser framework <parser-framework>`.
+Detect, parse, and change git URLs using libvcs's URL parser for `git(1)`. It builds on top of the
+{ref}`VCS-friendly URL parser framework <parser-framework>`.
 
 ```{eval-rst}
-.. automodule:: libvcs.parse.git
+.. automodule:: libvcs.url.git
    :members:
    :undoc-members:
    :show-inheritance:
