+++ src/libvcs/url/hg.py
@@ -6,8 +6,8 @@
   compare to :class:`urllib.parse.ParseResult`
 
   - Output ``hg(1)`` URL: :meth:`HgURL.to_url()`
-- Extendable via :class:`~libvcs.parse.base.MatcherRegistry`,
-  :class:`~libvcs.parse.base.Matcher`
+- Extendable via :class:`~libvcs.url.base.MatcherRegistry`,
+  :class:`~libvcs.url.base.Matcher`
 
 .. Note::
 
@@ -130,7 +130,7 @@ class HgURL(URLProtocol, SkipDefaultFieldsReprMixin):
     Attributes
     ----------
     matcher : str
-        name of the :class:`~libvcs.parse.base.Matcher`
+        name of the :class:`~libvcs.url.base.Matcher`
 
     Examples
     --------
